LayoutTests/platform/mac-wk1/imported/w3c/web-platform-tests/html/browsers/the-window-object/open-close/open-features-is-popup-condition_single-1-expected.txt


FAIL undefined should set BarProp visibility to true assert_equals: window.locationbar.visible expected true but got false
FAIL "popup" should set BarProp visibility to false assert_equals: window.menubar.visible expected false but got true
FAIL "popup=1" should set BarProp visibility to false assert_equals: window.menubar.visible expected false but got true
FAIL "popup=true" should set BarProp visibility to false assert_equals: window.menubar.visible expected false but got true
FAIL "popup=0" should set BarProp visibility to true assert_equals: window.locationbar.visible expected true but got false
FAIL "location" should set BarProp visibility to false assert_equals: window.menubar.visible expected false but got true
FAIL "location=yes" should set BarProp visibility to false assert_equals: window.menubar.visible expected false but got true
FAIL "location=true" should set BarProp visibility to false assert_equals: window.menubar.visible expected false but got true
FAIL "location=no" should set BarProp visibility to false assert_equals: window.menubar.visible expected false but got true
FAIL "toolbar" should set BarProp visibility to false assert_equals: window.menubar.visible expected false but got true
FAIL "toolbar=yes" should set BarProp visibility to false assert_equals: window.menubar.visible expected false but got true
FAIL "toolbar=true" should set BarProp visibility to false assert_equals: window.menubar.visible expected false but got true
FAIL "toolbar=no" should set BarProp visibility to false assert_equals: window.menubar.visible expected false but got true
FAIL "menubar" should set BarProp visibility to false assert_equals: window.menubar.visible expected false but got true
FAIL "menubar=yes" should set BarProp visibility to false assert_equals: window.menubar.visible expected false but got true
FAIL "menubar=true" should set BarProp visibility to false assert_equals: window.menubar.visible expected false but got true
FAIL "menubar=no" should set BarProp visibility to false assert_equals: window.menubar.visible expected false but got true
FAIL "resizable" should set BarProp visibility to false assert_equals: window.menubar.visible expected false but got true
FAIL "resizable=yes" should set BarProp visibility to false assert_equals: window.menubar.visible expected false but got true
FAIL "resizable=true" should set BarProp visibility to false assert_equals: window.menubar.visible expected false but got true
FAIL "resizable=no" should set BarProp visibility to false assert_equals: window.menubar.visible expected false but got true
